requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.10.5
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
distro==1.9.0
docstring_parser==0.16
et_xmlfile==2.0.0
firebase-admin==6.8.0
Flask==3.1.1
flask-cors==6.0.0
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.94.0
google-cloud-bigquery==3.33.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.16.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
lxml==5.4.0
MarkupSafe==3.0.2
msgpack==1.1.0
numpy==2.2.6
openai==1.82.0
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pinecone==7.0.1
pinecone-plugin-interface==0.0.7
proto-plus==1.26.1
protobuf==4.25.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-pptx==1.0.2
pytz==2025.2
requests==2.32.3
rsa==4.9.1
sentence-transformers==3.3.1
rank-bm25==0.2.2
scikit-learn==1.5.2
nltk==3.9.1
setuptools==80.8.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
waitress==3.0.2
websockets==15.0.1
Werkzeug==3.1.3
xlrd==2.0.2
XlsxWriter==3.2.3

# Phase 1: Core Infrastructure Dependencies
redis==5.0.1
upstash-redis==0.15.0
celery==5.3.4
flower==2.0.1
kombu==5.3.4

# LlamaIndex + NeonDB Dependencies
llama-index>=0.10.0
llama-index-vector-stores-postgres
llama-index-embeddings-huggingface
llama-index-llms-huggingface
llama-index-readers-file
llama-cloud-services==0.6.76
unstructured[pdf]>=0.15.15
python-magic>=0.4.27
asyncpg>=0.29.0
sqlalchemy>=2.0.0
pgvector>=0.3.0
requests>=2.31.0
gradio_client>=0.7.0

